Move UserUpdated to event subscribers
